Chapt. 13: i18n and l10n
This chapter was a slog as there was a lot to edit, these edits weren't documented
well and for me there was little payoff.
I didn't test this other than to make sure it ran.
  It's good to know how this mechanism works but it really slowed me down.
I kinda wish this was done after the reorg in ch 15 because you touch so many parts of the code.